+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   * [Added] tree option `.treeId` to prevent generation of a new sequence if the tree is re-initialized on a page.
   * [Changed] `.getTree()` now also accepts the tree id string
   * [Changed] #939: Keep a `partsel` flag that was explicitly set on a lazy node
+  * [Changed] ext-clones: make default key generation more robust against collisions
   * [DEPRECATED] loaderror and lazyload options now throw an error instead of falling back to the correct loadError and lazyLoad
   * [DEPRECATED] `tree.applyFilter` was removed
   * [Fixed] #918 SVG font awesome 5 glyphs remove badge counter when parent node is collapsed